simplify oauth
