+++ app/page.tsx
@@ -3,7 +3,7 @@ import { redirect } from 'next/navigation';
 import { Hero } from '@/components/home/hero';
 import { Features } from '@/components/home/features';
 import { HowItWorks } from '@/components/home/how-it-works';
-import PricingGrid from '@/components/idiot/auth/stripe/PricingTable';
+import PricingGrid from '@/components/idiot/auth/PricingTable';
 import { Contact } from '@/components/home/contact';
 import { CallToAction } from '@/components/home/call-to-action';
 import { ThemeToggle } from '@/components/layout/themes';
