+++ components/idiot/auth/Identify.tsx
@@ -8,8 +8,8 @@ import { z } from 'zod';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { setCookie } from 'cookies-next';
 import { LuUser } from 'react-icons/lu';
-import OAuth from '@/components/idiot/auth/oauth2/OAuth';
-import oAuth2Providers from '@/components/idiot/auth/oauth2/OAuthProviders';
+import OAuth from '@/components/idiot/auth/OAuth';
+import { providers as oAuth2Providers } from '@/components/idiot/auth/OAuth';
 import { useAuthentication } from '@/components/idiot/auth/Router';
 import AuthCard from '@/components/idiot/auth/AuthCard';
 import { Separator } from '@/components/ui/separator';
