+++ components/idiot/auth/PricingTable.tsx
