+++ components/idiot/auth/Subscribe.tsx
@@ -3,7 +3,7 @@
 import { getCookie } from 'cookies-next';
 import React, { Suspense } from 'react';
 import { useAuthentication } from '@/components/idiot/auth/Router';
-import PricingTable from '@/components/idiot/auth/stripe/PricingTable';
+import PricingTable from '@/components/idiot/auth/PricingTable';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import Link from 'next/link';
