--- components/idiot/auth/oauth2/Close.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import React, { useEffect } from 'react';
-import { useAuthentication } from '@/components/idiot/auth/Router';
-
-export type CloseProps = {};
-
-export default function Close() {
-  const authConfig = useAuthentication();
-
-  useEffect(() => {
-    window.close();
-  }, []);
-
-  return authConfig.close.heading ? <h2 className='text-3xl'>{authConfig.close.heading}</h2> : null;
-}
